Added a python version of the code and separated the C and Python code into their own directories
